web_services_classification/requirements.txt
# Core dependencies
absl-py==2.3.1
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dm-tree==0.1.9

# Core ML/Data Science
pandas==2.0.3
numpy==1.26.4
scikit-learn==1.3.0
scipy==1.15.3

# Deep Learning
tensorflow==2.15.0
keras>=2.15.0
torch 
torchvision

# NLP and Transformers
nltk==3.8
sentence-transformers==5.1.0
transformers==4.56.1
tokenizers==0.22.0
huggingface-hub==0.34.4
#textblob==0.19.0

# DeepSeek and PEFT Dependencies
peft>=0.15.0
bitsandbytes>=0.46.0
accelerate>=1.7.0
datasets>=3.6.0

# XGBoost
xgboost==1.7.6

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==6.3.0
wordcloud==1.9.4

# Configuration and Data Handling
PyYAML==6.0.1
joblib==1.3.2

# Utilities
tqdm>=4.66.3
requests==2.32.5
fsspec==2025.9.0
filelock==3.19.1

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Supporting libraries
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
kiwisolver==1.4.9
lark==1.2.2
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.2
ml-dtypes==0.2.0
namex==0.1.0
narwhals==2.4.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.9.1
rich==14.1.0
safetensors==0.6.2
six==1.17.0
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
